[bazel] Solving little warnings in the SW
//sw/device/silicon_creator/lib:verilator_boot_data_functest is flaky
with a long timeout so I've adjusted it to eternal to make it repeatable
and to prevent running it when we're under time pressure

The entropy_smoketest and the verilator_watchdog_functest are flaky with a
default timeout so I've increased them to long (900s)

I ran `bazel run //:buildifier_fix so it will be cleaner when run on
other systems and commits.

Also addressing the following warnings that appear during a SW build:

sw/device/lib/dif/dif_kmac_unittest.cc:115:23: warning: comparison of
integer expressions of different signedness: 'int' and 'const size_t'
{aka 'const long unsigned int'} [-Wsign-compare]
  115 |     for (int i = 0; i < size; ++i) {
      |                     ~~^~~~~~

sw/device/lib/dif/dif_kmac_unittest.cc:161:21: warning:
comparison of integer expressions of different signedness: 'int'
and 'long unsigned int' [-Wsign-compare]
  161 |   for (int i = 0; i < sizeof(uint32_t); i++) {
      |                   ~~^~~~~~~~~~~~~~~~~~

+++ sw/device/lib/dif/dif_kmac_unittest.cc
@@ -112,7 +112,7 @@ class KmacTest : public testing::Test, public mock_mmio::MmioTest {
    * @param size Len of the buffer.
    */
   void setExpectedMessageByte(const uint8_t *message, const size_t size) {
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
       EXPECT_WRITE8(KMAC_MSG_FIFO_REG_OFFSET, message[i]);
     }
   }
@@ -158,7 +158,7 @@ class AbsorbalignmentMessage : public KmacTest {};
 TEST_F(AbsorbalignmentMessage, Success) {
   uint8_t buffer[kMsg_.size() + sizeof(uint32_t)];
 
-  for (int i = 0; i < sizeof(uint32_t); i++) {
+  for (size_t i = 0; i < sizeof(uint32_t); i++) {
     uint8_t *pMsg = &buffer[i];
     std::copy(kMsg_.begin(), kMsg_.end(), pMsg);
